ms-learn/pyproject.toml
[project]
name = "ms-learn-agent"
version = "0.1.0"
description = "Convert Microsoft Learn courses to Word documents and audio"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    "edge-tts>=7.0.0",
    "click>=8.1.0",
]

[project.scripts]
mslearn = "mslearn.cli:cli"

[tool.setuptools.packages.find]
include = ["mslearn*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
